requirements.txt
dash==2.5.1
pandas==1.3.5
numpy==1.23.0
gunicorn==20.1.0
plotly==5.9.0
folium==0.12.1.post1
branca==0.5.0
python-dateutil==2.8.2
statsmodels==0.13.0
scikit-learn==1.1.0
skforecast==0.5.1
xgboost==1.6.2